enforce int type for slider values (#77)
Fix for Python 3.10 and PyQt 5.15 (possibly earlier versions too).

Required for the ROS-O project.

+++ joint_state_publisher_gui/src/joint_state_publisher_gui/__init__.py
@@ -100,7 +100,7 @@ class JointStatePublisherGui(QWidget):
 
             slider.setFont(font)
             slider.setRange(0, RANGE)
-            slider.setValue(RANGE/2)
+            slider.setValue(int(RANGE/2))
 
             joint_layout.addWidget(slider)
 
@@ -222,7 +222,7 @@ class JointStatePublisherGui(QWidget):
         self.update_sliders()
 
     def valueToSlider(self, value, joint):
-        return (value - joint['min']) * float(RANGE) / (joint['max'] - joint['min'])
+        return int((value - joint['min']) * float(RANGE) / (joint['max'] - joint['min']))
 
     def sliderToValue(self, slider, joint):
         pctvalue = slider / float(RANGE)
